Fix safepoint_recurse! with binary_trees example
It compiles, although I'm not sure if the macro is a
currently a completely safe abstraction.
We'll need to get compiletest working to verify that.

Safepoints/potential collections are now done *after* the clousre.
This makes more sense, because you would want to clean up garbage
after a sub-task (not before it).

There are runtime bugs right now, but that's fine :)