[simple] Use a linked list to store allocated objects
This is only working with the debug allocator right now.
This should actually save memory, since we use thin-pointers
instead of fat pointers and we avoid the overhead of a vec.

This slightly increses binary_trees time (n=21) from 1:04 to 1:07.
However it's probably worth the memory savings
and it'll be nessicarry for some planned changes to
the compacting collector ^_^