[copying] Actually apply the relocations to the pointer -_-
Right now at binary_trees (n=18), we have pretty much the same
performance as the simple collector. However, half the time
is spent in the system.... I suspect the problem is that
we're reallocating every collection.......

We produce the correct output for the sub trees,
but crash right before printing `long_lived_tree`.