Require double-indirection for values at all safepoints
The copying collcetor was passing `Gc` to the basic_safepoint
instead of `&mut Gc` which is used with recurse_safepoint!.

For debug reasons, the copying collector prints a backtrace
the first time a type is seen at a safepoint.

At first this appeared to fix the SEGFAULT at the end of the copying
example in binary_trees, but now it doesn't.