[simple] Inline type info into the object's header
This refactoring is nessicarry since I plan to implement a
custom segregated-fits allocator for small types.

Surprisingly, there is also a noticable performance improvement.
This appears to approve performance on binary_trees (n=21).
Previously this took 74 s + 513 MB,
now it's down to 57 s and 513 MB.

According to perf, we spend about 66% of time in `perform_collection`.
We spend 17-31% of our time in malloc and 12% in free.
Not quite sure how to read these reports, malloc is split into
a bunch of different functions.

Should consider lazy sweeping as defined in my GC Handbook.