[simple] While tracing, use recursion instead of an explicit stack
Apparently an explicit stack has a ton of performance overhead.
This is already what the copying collector,
so now we have much *fairer comparison numbers*.

This risks stack overflow for long chains of objects.
This was already a potential bug in the copying collector!!!!

At this point mark/sweep runs binary_trees (n=21) in 28 s
This is only 5 seconds slower (than the copying collector (23 s).

According to perf report, we spend 30% of our time tracing.
We are 18% slower than the copying collector, so lazy tracing
could definitely close the remaining gap :)