Use a GcRef trait in the API instead of a specific GC type
This gives implementations much more flexibility.
Fixes #5

DuckLogic is definitely going to use a custom pointer type,
so this is nessicarry for using zerogc for our collector.

On the downside this means duplicating common code in a lot of crates.
The simple `Gc` type has to be reimplemented in every collector.
If we ever add another basic collector to this crate, we should probably refractor out common code.
See also #4

For some reason this results in a couple seconds decrease in the performance of binary_Trees.
I'm assuming it's just some random noise,