Procedurally derive Trace and GcBrand.
Very close to #2

This significantly reduces the amount of unsafe code in user modules.
The only trait we don't auto-derive is `GcSafe`. This is a marker
trait so it's not too hard to implement.

We still aren't finished until we add support that.
We don't know if the type has an unsafely implemented destructor.
We could add our own no-op Drop implementation, but that would
break copy types. I'm thinking detecting `derive(Copy)` and allowing
the user to override if nessicarry.