First attempt at supporting write barriers: #3
This sets up the basic interface. All writes **must** notify
the owning reference with the field offset and the .
It's generic over the implementation, so Rust will create
specialized code for each type/GC combination.
The default is a no-op and is #[inline(always)].
Write barriers (and any other mutation) is unsafe,
since they're relied upon for correctness.

We're relying on procedural derive to generate safe setter methods
that wrap the underling `GcCell` and invoke the write .
Unfortunately, the current implementation requires that
all mutable fields have the type `GcCell<Gc<'gc, Value>>`.
This is way to restrictive, even for our simple binary_trees example.

Could we somehow make some sort of `WriteBarrier` trait?