[simple] Attempt to support freezing simple collectors
This is completely untested.
TODO: Undefined behavior to drop (or otherwise forget to unfree) a
frozen collector.

Maintaining two seperate codebases for sync/not-sync is horrible.
I think we should just drop support for single-threaded code......

On the other hand, depending on crossbeam + parking_lot pulls in
a ton of dependencies.