Use opt-level=1 in debug mode
I wasn't catching bugs because everything took way too long.
Now we're finally getting corruption of Gc::collector_ptr

Looks like we spend way too long waiting for locks so
we aren't ever beginning collections.
I'm betting this is be because of add_context/free_context

Maybe we should use a RwLock. We really just acquire the lock
in add_context to hold off any collections, essentially making it a
read lock.
This means that readers trying to acquire the lock will
block even if the lock is unlocked when there are writers waiting to acquire the lock.