[simple] Compile without optimizations, revealing deadlock again
Good news is this fixes the collections taking forever to happen.
Now they actually free memory. Unfortunately its still not as
good as the single threaded collector (probably still due to locking issues).

It appears the locks in add_context/free_context are starving
the lock in trigger_context. Maybe we can prevent this by switching
to a RwLock? I gave up getting Condvar working before but
it looks like there is a workaround:
https://github.com/Amanieu/parking_lot/issues/165