[simple::context] Use RwLock to protect CollectorState instead of a plain Mutex
This creates a hard dependency on parking lot, since we now rely on the behavior
of their RwLock. The good news is it should fix lock starvation in `trigger_safepoint` :)

The new problem is that a new context is being allocated at the same address as the old one
and for some reason the old one isn't registered as being freed (causing duplicates).
Why would the new add_context ever execute before the old free_context has dropped the
original memory for the context? The allocator should ensure that only one address is in use.....