[simple::context] Repeatedly check for valid_contexts.len() == total_contexts inside of await_collection
This moves the responsibility from detecting the start of a collection from
`push_pending_context` over to await_collection.

The could be triggered by freeing a context, not just a call to push_context.
Therefore we need to loop the check repeatedly!
The binary_trees_parallel example finaly works in debug mode!
However in release it SEGFAULTs inside visit_gc :(
It looks like there's some sort of invalid root?

Maybe it has to do with frozen contexts.
I haven't really tested that since it's only used in the parallel collector.