[simple] Correctly trace GcHandleList. Actually tracing pointers as `Gc<T>` instead of `T`
Required refractoring the gc tracing code into MarkVistor::visit_raw_gc.

Marked the `GcHandle` wrapper as `NullTrace` as the value it points to should be a root.
It's tracked seperately and doesn't need to be traced. This should probably be documented
somewhere.....

As of this commit, both debug and release builds of the parallel collector
consistently pass valgrind and produce the correct output.
Fixes #8

We should almost be ready for use in DuckLogic :O