Add a default feature-flag for "multiple-collectors"
If this is off, than `SimpleCollector` is a global singleton.
Gc pointers don't need to store a reference to their owning collector,
so they should be just a single thin pointer.

Make `Gc` pointer repr(C), since DuckLogic will rely on the assumption
it's represented just like a simple pointer.
Outside of safepoints, that's all it really is ;)