[context] Use traits instead of cfg! to handle singleton collectors
Wheter the **implementation** of zerogc-simple is a singleton is still
controlled by features,
but the public API of the context crate is not.
This is the first step on the way towards making the implemnentation more configurable.

Now whether or not a collector is a singleton is handled by
if they implement 'SingletonCollector' and if 'RawSimpleCollector::Singleton'
is true/false.

Now we can check compilation for both multiple-collectors and singleton
at the same time :)

This way we can have some collectors that are singletons and
some that are not within the same compilation units.
With the old feature-based approach, making one collector have
multiple instances forced all of them too.
Eventually, we need to support sync/nosync the same way.