[BREAKING] Split GcBrand up into GcErase + GcRebrand
This helps clarify its two seperate purposes.
We stil require T: NullTrace for the `&'a T` implementation.
This seems like a requirement for the time beeing :(

Give warnings for the auto-derived implementations of GcRebrand and
GcErase indicating that we (currently) don't check the correctness
of the implementation on each of the fields.

I've changed the double-indirection and '&mut &mut T' behavior I
had setup back when we used a copying collector. This doesn't break anything
right now (because we use a non-moving collector).
However, I have to be really careful if we ever switch back!

This commit is so incompatible it will force us towards a 0.2.0-alpha release.