Support deriving erase/rebrand for non-gc lifetimes if Self: NullTrace
Emits a warning to reflect the current unsafety/uncheckedness of
the GcErase/GcRebrand default-implementations.
TODO: Fix this

This is due to current limitations in the type-system and
how we choose to implement GcErase + GcRebrand.
See the docs of `&'a T` for more on why we require `T: NullTrace`
TODO: Figure out a way around this.