[derive] Implement safety checks for GcErase + GCRebrand
Remove those nasty warnings

Add a #[zerogc(collector_id(Id))] to mark which specific collector
you support rebranding for. If you have a `Gc<'gc, SpecificId>` than
It's not safe to implement a general `for<I: CollectorId> GcErase<'a, I>`
because `I` might not match `SpecificId`.

This was caught with the new verification! Horay for type safety!