Allow customizing whether GcArray is fat or thin
Switch the epsilon collector to use fat pointers for arrays,
so that it's easy to transmute back and forth to `&[T]`.
This allows us to remove that horrible `epsilon_static_array!` macro.

Futher lowers the overhead of the epsilon collector......

TODO: We should be able to eliminate most array headers
thanks to this......