`T: GcSafe<'gc, T>` should no longer imply `T: 'gc` (#32)
Clarify the lifetime and ownership for `Gc`.
A Gc pointer is logically equivalent to a pair of (*mut T, &'gc Id::System). It does *not* contain a &'gc T like you would expect.
This means that we can weaken the `T: 'gc` lifetime, which paves the way for temporary borrows that are shorter than the 'gc lifetime. In other words, it should eventually be possible to have &'a T even if 'a < 'gc

Also, we can safely make the lifetime covariant now :)

Merges #32 

I had tried to make GcRebrand take advantage of generic associated types, but that was an utter failure.