Add a GcError wrapper
Allows garbage collected errors that are otherwise `Trace`,
to be impl `std::error::Error + 'static`.

This uses a `GcHandle` to convert all 'gc lifetimes to 'static.

The really neat thing about this is that it's implemented
entirely in safe code, so you can define similar compatibility
abstractions for your own code.