Fix lifetime covariance for GcArray
This is fixed by making `Id::ArrayPtr` untyped as `NonNull<c_void>`
instead of `NonNull<T>`. Before this, the borrow checker couldn't "see through" the
generics, and has to assume 'T' is invariant'

See nomicon for details on variance:
https://doc.rust-lang.org/nomicon/subtyping.html#variance

Add a test to the epsilon collector to make sure this doesn't regress.