Add GcHeader trait & GcContext type
The GcHeader api allows access to an object's internal header,
which contains the CollectorId.
For an array, the header also contains the length.

Implmentations can (and will) add extra metadata to their header
like mark bits and type info.

Make methods on Gc associated functions instead of methods
in order to avoid conflicts with the Deref trait.
This is what Arc and Box both do.